Fix bug with float pixel
e.g. screen width is 1903 and the elements have width 50%, therefore 951.5px. OffsetWidth is the uprounded value (952) and therefore the check for possible visible elements fails (comparison of 2 elements with 50% width and the screen width does not result that there is enough space for two elements)

Also prevents unnecessary reflows by putting the result of getComputedStyle in a variable